+++ INSTALL.md
@@ -8,9 +8,9 @@ This page covers base installation and optional dependencies (“extras”).
 pip install scikit-fingerprints
 ```
 
-## Optional dependencies (extras)
+## Optional dependencies
 
-Some functionality depends on optional dependencies. Install them via extras:
+Some functionalities depend on optional dependencies. They are installed as package "extras", as shown below.
 
 ### Neural fingerprints (CLAMP): `neural`
 
@@ -20,8 +20,7 @@ pip install "scikit-fingerprints[neural]"
 
 This installs **PyTorch** (`torch`).
 
-If you need a specific PyTorch build (CPU/CUDA), follow the official PyTorch installation instructions
-first, then install scikit-fingerprints.
+If you need a specific PyTorch build (CPU/CUDA), install PyTorch following the official instructions. We highly recommend using [uv](https://docs.astral.sh/uv/) for this to avoid dependency conflicts.
 
 For example, CPU wheels via the official PyTorch index:
